core.info
